requirements.txt
langchain[all]
openai
python-docx
